# docs/swagger.yaml
basePath: /api/v1
definitions:
  fiber.Error:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
  fiber.Map:
    additionalProperties: true
    type: object
  models.SendNotificationRequest:
    properties:
      message:
        type: string
      meta_data:
        additionalProperties: true
        description: Optional metadata for additional information
        type: object
      recipient:
        type: string
      title:
        type: string
      type:
        type: string
    type: object
  models.SendNotificationResponse:
    properties:
      id:
        type: string
      status:
        type: string
    type: object
host: localhost:3000
info:
  contact:
    email: ngdangkietswe@yopmail.com
    name: API Support
  description: This is a notification service API with RabbitMQ integration
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Notification Service API
  version: "1.0"
paths:
  /api/v1/health:
    get:
      description: Check if the notification service is running
      produces:
      - application/json
      responses:
        "200":
          description: Service is running
          schema:
            $ref: '#/definitions/fiber.Map'
      summary: Health check
      tags:
      - Health
  /api/v1/notifications:
    post:
      consumes:
      - application/json
      description: Send a notification to a recipient
      parameters:
      - description: Notification request
        in: body
        name: notification
        required: true
        schema:
          $ref: '#/definitions/models.SendNotificationRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/models.SendNotificationResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/fiber.Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/fiber.Error'
      summary: Send a notification
      tags:
      - Notifications
  /api/v1/rabbitmq/bindings:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all RabbitMQ bindings
      produces:
      - application/json
      responses:
        "200":
          description: List of RabbitMQ bindings
          schema:
            items:
              type: object
            type: array
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/fiber.Error'
      summary: Get list of RabbitMQ bindings
      tags:
      - RabbitMQ
  /api/v1/rabbitmq/exchanges:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all RabbitMQ exchanges
      produces:
      - application/json
      responses:
        "200":
          description: List of RabbitMQ exchanges
          schema:
            items:
              type: object
            type: array
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/fiber.Error'
      summary: Get list of RabbitMQ exchanges
      tags:
      - RabbitMQ
  /api/v1/rabbitmq/queues:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all RabbitMQ queues
      produces:
      - application/json
      responses:
        "200":
          description: List of RabbitMQ queues
          schema:
            items:
              type: object
            type: array
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/fiber.Error'
      summary: Get list of RabbitMQ queues
      tags:
      - RabbitMQ
schemes:
- http
- https
swagger: "2.0"